refactor: extract THEMES to shared constants and add invalid-key guard test
