+++ src/components/themes.ts
@@ -1,0 +1,8 @@
+export const THEMES = [
+  { key: "",       name: "Ambre",  bg: "#fffbf0", sidebar: "#1e293b", accent: "#f59e0b" },
+  { key: "ocean",  name: "Ocean",  bg: "#eff6ff", sidebar: "#0c1a2e", accent: "#06b6d4" },
+  { key: "forest", name: "Forest", bg: "#f0fdf4", sidebar: "#052e16", accent: "#84cc16" },
+  { key: "sunset", name: "Sunset", bg: "#fff7ed", sidebar: "#431407", accent: "#f97316" },
+  { key: "galaxy", name: "Galaxy", bg: "#faf5ff", sidebar: "#1a1030", accent: "#8b5cf6" },
+  { key: "rose",   name: "Rose",   bg: "#fff1f2", sidebar: "#1f0613", accent: "#f43f5e" },
+] as const;
